feat: post controllers done

+++ server/controllers/postController.js
@@ -3,7 +3,7 @@ const Post = require('../models/PostModels')
 const createPost = async (req,res)=>{
   try {
         const { content, media, tags } = req.body
-        if (!media.trim()) return res.status(500).json({ message: "Media fields is required "})
+        if (!media?.trim()) return res.status(400).json({ message: "Media fields is required "})
         const user = req.user
         const newPost = await Post.create({
             userId: user._id,
@@ -12,8 +12,9 @@ const createPost = async (req,res)=>{
             tags
     })
 
-        if (!newPost) return res.status(500).json({sucess: false, message: "Error creating post "})
-        res.status(201).json({ sucess: true, message: "new post successfully created "})
+        if (!newPost) return res.status(500).json({success: false, message: "Error creating post "})
+
+        res.status(201).json({ success: true, message: "New post successfully created "})
   } catch (error) {
         console.error(error.message)
         res.status(500).json({ message: error.stack})
@@ -24,13 +25,14 @@ const createPost = async (req,res)=>{
 const getAllPost = async (req, res) =>{
     try {
         const posts = await Post.find()
-                        .populate("user","fullName userName bio likes")
+                        .populate("userId","fullName userName profilePicture")
                 
-        if (posts.length == 0)return res.status(404).json({ message: "No available post"})
-        res.status(200).json({sucess: true})
+        if (posts.length == 0)return res.status(200).json({ message: "No available posts", posts:[],})
+
+        res.status(200).json({success: true, posts: posts})
     }catch (error) {
         console.error(error.message)
-        res.status(500).json({ sucess: false})
+        res.status(500).json({ success: false, message: " internal Server Error "})
     }
 
 }
@@ -39,21 +41,34 @@ const getSinglePost = async (req, res)=>{
     try {
         const { id } = req.params
         const post = await Post.findById(id)
-        if (!post) return res.status(400).json({ message: "Post not found"})
+                        .populate("userId", "fullName userName profilePicture")
+                        .populate("comments.user", "fullName userName profilePicture")
+
+        if (!post) return res.status(404).json({ message: "Post not found"})
     
-    res.status(200).json({ sucess: true})
+    res.status(200).json({ success: true, post: post})
     } catch (error) {
         console.error("An error has occured: ",error.message)
+        res.status(500).json({ message: "internal server error "})
     }
 }
 
 const updatePost = async (req, res)=>{
    try {
     const { id } = req.params
+    const user = req.user
+
+    //check if post exists and belongs to a user
+    const postIsExists = await Post.findById(id)
+    if (!postIsExists) return res.status(404).json({ success: false, message: "Post not found "})
+    
+    if (postIsExists.userId.toString() !== user._id.toString()) return res.status(403).json({ message: "you can only update your own posts"})
+
     const updatedPost = await Post.findByIdAndUpdate(id, req.body, {runValidators: true, new: true})
+                    .populate("userId", "fullName userName profilePicture")
     if (!updatedPost) return res.status(500).json({ message: "update failed"})
     
-    res.status(200).json({ sucess: true, updatedPost})
+    res.status(200).json({ success: true, post: updatedPost,})
    } catch (error) {
     console.error("Error updating Post:" + error.message)
     res.status(500).json({message: "Error updating post try again!!!"})
@@ -63,15 +78,26 @@ const updatePost = async (req, res)=>{
 
 const deletePost = async (req, res) =>{
     const { id } = req.params
-    const deletedPost = await Post.findByIdAndDelete(id, req.body, {new: true})
-    if (!deletedPost) return res.status(500).json({ message: "unable to delete post "})
+    const user = req.user
 
-    res.status(200).json({sucess: true, message: "sucessfully deleted"} )
+    try {
+        const postIsExists = await Post.findById(id)
+        if (!postIsExists) return res.status(404).json({success: true, message: "Post Not Found "})
+
+        if (postIsExists.userId.toString() !== user._id.toString()) return res.status(403).json({ success: false, message: "you can only delete your own post"})
     
+        const deletedPost = await Post.findByIdAndDelete(id, req.body, {new: true})
+        if (!deletedPost) return res.status(500).json({ message: "unable to delete post "})
+
+        res.status(200).json({success: true, message: "successfully deleted"} )
+    } catch (error) {
+       console.error("Delete post Error: " + error.message )
+       res.status(500).json({ success: false, message: "internal Server Error "}) 
+    }
     
 }
 
-const togglelike = async () =>{
+const toggleLike = async (req, res) =>{
   try {
         const { id } = req.params
         const user = req.user
@@ -81,17 +107,30 @@ const togglelike = async () =>{
 
         const likedpost = post.likes.includes(user._id)
         if (likedpost){  //if already liked
-            post.likes.pull(user._id)  //post.like.filter(id => id.toString() !=== _id.toString())
+            post.likes = post.likes.filter(id => id.toString() !== user._id.toString())
     }else{
-            post.likes.push(user._id)
+        post.likes.push(user._id)
     }
-    post.save()
-    res.status(200).json({
-        sucess: true,
-        message: likedpost ? "post unliked" : "post liked",
-        likedCount: post.likes.length
+        post.save()
+
+        res.status(200).json({
+            success: true,
+            message: likedpost ? "post unliked" : "post liked",
+            likedCount: post.likes.length,
+            likedpost: !likedpost
+            
     })
   } catch (error) {
-    console.error("unable to like post: " + error.qmessage)
+        console.error("unable to like post: " + error.message)
+        res.status(500).json({ message: "internal server Error "})
   }
+}
+
+module.exports = {
+    createPost,
+    getAllPost,
+    getSinglePost,
+    updatePost,
+    deletePost,
+    toggleLike
 }
